Bug 780279 - Expose SyncAccounts.requestImmediateSync().
