+++ src/main/java/org/mozilla/gecko/sync/setup/SyncAccounts.java
@@ -61,6 +61,56 @@ public class SyncAccounts {
   }
 
   /**
+   * Asynchronously request an immediate sync, optionally syncing only the given
+   * named stages.
+   * <p>
+   * Returns immediately.
+   *
+   * @param account
+   *          the Android <code>Account</code> instance to sync.
+   * @param stageNames
+   *          stage names to sync, or <code>null</code> to sync all known stages.
+   */
+  public static void requestImmediateSync(final Account account, final String[] stageNames) {
+    if (account == null) {
+      Logger.warn(LOG_TAG, "Not requesting immediate sync because Android Account is null.");
+      return;
+    }
+
+    final Bundle extras = new Bundle();
+    Utils.putStageNamesToSync(extras, stageNames, null);
+    extras.putBoolean(ContentResolver.SYNC_EXTRAS_MANUAL, true);
+    ContentResolver.requestSync(account, BrowserContract.AUTHORITY, extras);
+  }
+
+  /**
+   * Asynchronously request an immediate sync for all Sync accounts, optionally
+   * syncing only the given named stages.
+   * <p>
+   * Returns immediately; safe to call from main thread.
+   *
+   * @param context
+   *          Android context.
+   * @param stageNames
+   *          stage names to sync, or <code>null</code> to sync all known stages.
+   */
+  public static void requestImmediateSync(final Context context, final String[] stageNames) {
+    new AsyncTask<Void, Void, Void>() {
+      @Override
+      protected Void doInBackground(Void... params) {
+        // The syncAccounts method should not be called on the
+        // UI thread as it touches disk to access a sqlite DB.
+        final Account[] accounts = syncAccounts(context);
+
+        for (Account account : accounts) {
+          requestImmediateSync(account, stageNames);
+        }
+        return null;
+      }
+    }.execute();
+  }
+
+  /**
    * Asynchronously invalidate the auth token for a Sync account.
    *
    * @param accountManager Android account manager.
